correct warnings

+++ src/phase/bline.c
@@ -1,6 +1,6 @@
 /*   File      : S_UF/afs/psi.ch/user/f/flechsig/phase/src/phase/bline.c */
 /*   Date      : <10 Feb 04 16:34:18 flechsig>  */
-/*   Time-stamp: <2013-06-29 11:33:05 flechsig>  */
+/*   Time-stamp: <02 Jul 13 08:26:22 flechsig>  */
 /*   Author    : Uwe Flechsig, flechsig@psi.ch */
  
 /*   $Source$  */
@@ -40,7 +40,7 @@ void debug_beamline_type_c_(int *ip)
 {
   struct BeamlineType *bl;
   bl= (struct BeamlineType *)ip;
-  printf("%s debug_beamline_type_c_: %d position: %d\n", __FILE__, ip, bl->position);
+  printf("%s debug_beamline_type_c_: %d position: %d\n", __FILE__, (int)ip, bl->position);
 } /* end debug_beamline_type_c_ */
